+++ styles/styles.css
@@ -188,7 +188,7 @@ em strong {
   font-weight: 700;
 }
 
-main.text-caption {
+.text-caption {
   margin: 15px 0 25px;
   font-size:var(--body-font-size-xs);
   display: block;
