+++ templates/application-note/application-note.css
@@ -1,0 +1,5 @@
+main img {
+  display: block;
+  margin: 0 auto;
+  border: 0;
+}
