+++ templates/config.js
@@ -1,5 +1,5 @@
 /**
  * to add/remove a template, just add/remove it in the list below
  */
-const TEMPLATE_LIST = ['resource'];
+const TEMPLATE_LIST = ['application-note', 'news', 'publication'];
 export default TEMPLATE_LIST;
