+++ templates/news/news.css
@@ -1,9 +1,3 @@
-main img {
-  display: block;
-  margin: 0 auto;
-  border: 0;
-}
-
 p {
   margin-top: 0;
   margin-bottom: 10px;
