+++ templates/publication/publication.css
@@ -1,0 +1,1 @@
+@import "../news/news.css";
